Half way through doing the mouse code.
Just need to add more variables and methods to the Input.h and link them up in WinGLApplication.cpp. By the end there should be mouseDown array which will give not only the buttons, but also if the scroll is up or down ( `Input::mouseDown[MOUSE_SCROLL_UP]` will be true for example).